+++ backapp/DAL/Repo/VacationRepo.cs
@@ -1,42 +1,47 @@
 ﻿using DAL.EF;
 using DAL.Interfaces;
+using Microsoft.EntityFrameworkCore;
 
 namespace DAL.Repo
 {
     internal class VacationRepo : Repo, IRepo<Vacation, int, Vacation>
     {
-        public Vacation Add(Vacation obj)
+        public async Task<Vacation?> Add(Vacation obj)
         {
-            db.Vacations.Add(obj);
-            if (db.SaveChanges() > 0)
-                return obj;
-            return null;
+            await db.Vacations.AddAsync(obj);
+            await db.SaveChangesAsync();
+            return obj;
         }
 
-        public bool Delete(int id)
+        public async Task<bool> Delete(int id)
         {
-            var dbpost = Get(id);
-            db.Vacations.Remove(dbpost);
-            return db.SaveChanges() > 0;
+            var dbpost = await Get(id);
+            if(dbpost != null)
+            {
+                db.Vacations.Remove(dbpost);
+            }
+            return await db.SaveChangesAsync() > 0;
         }
 
-        public List<Vacation> Get()
+        public async Task<List<Vacation>?> Get()
         {
-            return db.Vacations.ToList();
+            return await db.Vacations.ToListAsync();
         }
 
-        public Vacation Get(int id)
+        public async Task<Vacation?> Get(int id)
         {
-            return db.Vacations.Find(id);
+            return await db.Vacations.FindAsync(id);
         }
 
-        public Vacation Update(Vacation obj)
+        public async Task<Vacation?> Update(Vacation obj)
         {
             var dbpost = Get(obj.Id);
-            db.Entry(dbpost).CurrentValues.SetValues(obj);
-            if (db.SaveChanges() > 0)
-                return obj;
-            return null;
+            if(dbpost != null)
+            {
+                db.Entry(dbpost).CurrentValues.SetValues(obj);
+            }
+            await db.SaveChangesAsync();
+            return obj;
         }
     }
 }
